S15.e.2: extract operator.ts sandbox-list fetchers (2739 → 2483 LOC)
Second sub-slice of the S15.e operator.ts decomposition train.

Lifts the three sandbox-list fetchers out of the giant startDashboard
closure into a sibling module:

- fetchSandboxes(kubeContext?) — unified Docker + AKS view
- fetchSandboxesAKS(kubeContext?) — kubectl-driven CRD + pod + secret
  inspection with parallel per-sandbox enrichment
- fetchSandboxesDocker() — docker ps + per-container env/handoff probe

→ cli/src/commands/operator/fetchers/sandboxes.ts (287 LOC)

Also moves the 2-line module-level kctl(args, context?) helper to
operator/helpers.ts so the new fetcher module can reuse it without
re-implementing the --context prefix logic.

Behavioral change: NONE.
- Bodies byte-identical to originals.
- kctl byte-identical: return context ? ["--context", context, ...args] : args
- Mechanical edit: closure-captured kubeContext now passed as explicit
  parameter; the single call site in startDashboard.refresh() updated
  to forward it. Tree ordering, handoff state detection, channel parsing,
  health-state derivation, secret-data scraping — all unchanged.

Cumulative S15.e delta:
  pre-S15.e: 2894
  e.1:       2739 (-155) — types.ts + helpers.ts
  e.2:       2483 (-256) — fetchers/sandboxes.ts (+ kctl move)
  Cap (800): not yet met; multi-PR sub-train continues.

Verification:
- operator.ts LOC: 2739 → 2483
- tsc --noEmit clean / lint 27 warnings (baseline) / build clean
- npm test: 454 pass / 2 skipped pre-existing

Removed unused HealthState import from operator.ts (now only used
inside the extracted fetcher module).

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e2.md
